requirements.txt
gspread
pandas
scikit-learn
nltk
google-auth
